+++ SerialPrograms/Source/Pokemon/Inference/Pokemon_BoxGenderDetector.h
@@ -24,14 +24,14 @@ public:
     // box: the area where the gender symbol will appear
     // area_ratio_threshold: if the number of red pixel count > box total area * area_ratio_threshold, female symbol is detected
     //                       if the number of blue pixel count > box total area * area_ratio_threshold, male symbol is detected
-    BoxGenderDetector(const ImageFloatBox& box, float area_ratio_threshold, Color color = COLOR_RED);
+    BoxGenderDetector(const ImageFloatBox& box, double area_ratio_threshold, Color color = COLOR_RED);
 
     void make_overlays(VideoOverlaySet& items) const;
     EggHatchGenderFilter detect(const ImageViewRGB32& screen) const;
 
 private:
     ImageFloatBox m_box;
-    float m_area_ratio_threshold;
+    double m_area_ratio_threshold;
     Color m_color;
 };
